# docs/_generated/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API'
  description: 'Headless content platform: entries, taxonomies, media, auth.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost/api/v1'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Получите JWT через <code>POST /api/v1/auth/login</code> или используйте refresh endpoint.'
security:
  -
    default: []
paths:
...
  '/api/v1/admin/options/{namespace}':
    get:
      summary: 'Список опций по namespace.'
      operationId: Namespace
      description: ''
      parameters:
        -
          in: query
          name: q
          description: 'Поиск по ключу/описанию (<=255).'
          example: hero
          required: false
          schema:
            type: string
            description: 'Поиск по ключу/описанию (<=255).'
            example: hero
            nullable: false
        -
          in: query
          name: deleted
          description: 'Управление soft-deleted. Values: with,only.'
          example: consequatur
          required: false
          schema:
            type: string
            description: 'Управление soft-deleted. Values: with,only.'
            example: consequatur
            nullable: false
        -
          in: query
          name: per_page
          description: 'Размер страницы (1-100). Default: 20.'
          example: 17
          required: false
          schema:
            type: integer
            description: 'Размер страницы (1-100). Default: 20.'
            example: 17
            nullable: false
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 10
                      namespace: site
                      key: hero.title
                      value: 'Launch the future'
                      description: 'Hero headline'
                      updated_at: '2025-01-10T12:00:00+00:00'
                      deleted_at: null
                  links:
                    first: …
                    last: …
                    prev: null
                    next: null
                  meta:
                    current_page: 1
                    last_page: 1
                    per_page: 20
                    total: 1
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 10
                        namespace: site
                        key: hero.title
                        value: 'Launch the future'
                        description: 'Hero headline'
                        updated_at: '2025-01-10T12:00:00+00:00'
                        deleted_at: null
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 10
                        namespace:
                          type: string
                          example: site
                        key:
                          type: string
                          example: hero.title
                        value:
                          type: string
                          example: 'Launch the future'
                        description:
                          type: string
                          example: 'Hero headline'
                        updated_at:
                          type: string
                          example: '2025-01-10T12:00:00+00:00'
                        deleted_at:
                          type: string
                          example: null
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                        example: …
                      last:
                        type: string
                        example: …
                      prev:
                        type: string
                        example: null
                      next:
                        type: string
                        example: null
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 20
                      total:
                        type: integer
                        example: 1
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  type: 'https://stupidcms.dev/problems/unauthorized'
                  title: Unauthorized
                  status: 401
                  detail: 'Authentication is required to access this resource.'
                properties:
                  type:
                    type: string
                    example: 'https://stupidcms.dev/problems/unauthorized'
                  title:
                    type: string
                    example: Unauthorized
                  status:
                    type: integer
                    example: 401
                  detail:
                    type: string
                    example: 'Authentication is required to access this resource.'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  type: 'https://stupidcms.dev/problems/invalid-option-identifier'
                  title: 'Validation error'
                  status: 422
                  detail: 'The provided option namespace/key is invalid.'
                  errors:
                    namespace:
                      - 'The namespace format is invalid.'
                properties:
                  type:
                    type: string
                    example: 'https://stupidcms.dev/problems/invalid-option-identifier'
                  title:
                    type: string
                    example: 'Validation error'
                  status:
                    type: integer
                    example: 422
                  detail:
                    type: string
                    example: 'The provided option namespace/key is invalid.'
                  errors:
                    type: object
                    properties:
                      namespace:
                        type: array
                        example:
                          - 'The namespace format is invalid.'
                        items:
                          type: string
        429:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Too Many Attempts.'
                properties:
                  message:
                    type: string
                    example: 'Too Many Attempts.'
      tags:
        - 'Admin ▸ Options'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                namespace:
                  type: string
                  description: 'Must match the regex /^[a-z0-9_][a-z0-9_.-]{1,63}$/.'
                  example: gvxi44f2f25h5.6eezattmmlhydv_tra-jle3y2y8
                  nullable: false
                q:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: bajwbpilpmufinllwloau
                  nullable: true
                deleted:
                  type: string
                  description: ''
                  example: with
                  nullable: true
                  enum:
                    - with
                    - only
                page:
                  type: integer
                  description: 'Must be at least 1.'
                  example: 87
                  nullable: true
                per_page:
                  type: integer
                  description: 'Must be at least 1. Must not be greater than 100.'
                  example: 4
                  nullable: true
              required:
                - namespace
    parameters:
      -
        in: path
        name: namespace
        description: 'Пространство опций (a-z0-9_.-).'
        example: site
        required: true
        schema:
          type: string
...